recommender_system/requirements.txt
matplotlib
numpy
pandas
scipy
sklearn
scikit-surprise


